back/mypacket/requirement.txt
django==2.1
djangorestframework==3.11.0
httpie==2.0.0
django-cors-headers==3.2.1
django_rest_passwordreset